refactor(graph): extract InternetHub to js/graph/internet-hub.js and import in app.js
